Add Item class and a Controller for managing Items
Add the data class Item with the fields from Firestore (firebaseId, name,
description, isShared, quantity, unit, sharedEmail, bestByDate, owner).

Add the Controller 'ItemController' that creates Items based on Firestore
documents and offers functionality to saveItems to Firestore.

Add the 'ItemControllerInterface' that defines an Interface for classes to
use when accessing and storing Items from/to Firestore.

Add unit tests for the 'Item' class that test the initialization of
Items with parameters that are null.

Add the function 'getByValue' to 'Unit' to make it possible to choose a
unit by an index.

Add the Firestore dependency to 'build.gradle'.