Add functionality to UI components in EditFragment
Add functionality to the save button for new Items to
store new Items temporarily.
Add functionality to the 'locate me' button to determine
the device's location with just a tap.
Add a DatePicker so that the user can conveniently pick
the best by date without having to type it in with the keyboard.
Grey out the location picker elements when the sharing option is
disabled.