Add further fields to objects of Item class
Add the following fields to 'Item' objects:
  - location: The Item's pick up location with both longitude
    and latitude
  - geohash: The locations hash code which enables coarse
    sorting of items by location
  - contactName: The owners name displayed, when viewing
    a shared item
  - contactEmail: The owners email displayed when viewing
    a shared item
  - ownerReference: The reference to the owner's document
    on Firestore

Support parsing for the new fields added to Item.

Adapt UnitTests to the new constructor of Item. Due to some
problems with Firebase, the UnitTests don't work anymore.
This has to be fixed in upcoming commits.

Add a dependency to build.gradle to get access to classes
needed for the Item's location field.