Refactor date related functions and Unit enum
Move functions related to Date and DatePicker calculations into
Utils file, so that the EditFragment becomes less bloated.

Provide each Unit with a symbolId that references the related
unit symbol as a resource string.

Add a function that returns a Unit to a given string.

Remove or adapt affected functions in EditFragment.