Implement deleteItem and add swipe support
Add trash icon to drawable folder.
Update the toast text which is display when a new Item is saved.
Fill 'deleteItem' method in ItemController with functionality.
Add a class that handles all 'swipe to delete' events on Item rows in
the pantry section ('ItemSwipeToDeleteCallback').
Integrate the new class into the PantryFragment.